+++ src/main/java/de/mari_023/ae2wtlib/wet/WETScreen.java
@@ -6,15 +6,24 @@ import net.minecraft.world.entity.player.Inventory;
 import appeng.client.gui.me.items.PatternEncodingTermScreen;
 import appeng.client.gui.style.ScreenStyle;
 
+import de.mari_023.ae2wtlib.terminal.ScrollingUpgradesPanel;
 import de.mari_023.ae2wtlib.terminal.WTMenuHost;
 import de.mari_023.ae2wtlib.wut.IUniversalTerminalCapable;
 
 public class WETScreen extends PatternEncodingTermScreen<WETMenu> implements IUniversalTerminalCapable {
+    private final ScrollingUpgradesPanel upgradesPanel;
+
     public WETScreen(WETMenu container, Inventory playerInventory, Component title, ScreenStyle style) {
         super(container, playerInventory, title, style);
         if (getMenu().isWUT())
             addToLeftToolbar(cycleTerminalButton());
-        addUpgradePanel(widgets, getMenu());
+        upgradesPanel = addUpgradePanel(widgets, getMenu());
+    }
+
+    @Override
+    public void init() {
+        super.init();
+        upgradesPanel.setMaxRows(Math.max(2, getVisibleRows()));
     }
 
     @Override